+++ src/docs/partials/change-log.html
@@ -1,6 +1,28 @@
 <page-body>
   <section>
     <h1>Version 6</h1>
+    <h2>Beta7</h2>
+    <h3>Bug Fixes</h3>
+    <ul>
+      <li>Fixed options mutable. <a
+              href='https://github.com/Eonasdan/tempus-dominus/issues/2487'>#2487</a></li>
+      <li>Fixed element toggle when input is disabled <a href='https://github.com/Eonasdan/tempus-dominus/issues/2495'>#2495</a></li>
+    </ul>
+    <h3>New</h3>
+    <ul>
+      <li>Added a new example for setting and getting dates.</li>
+    </ul>
+    <h2>Beta6</h2>
+    <h3>Bug Fixes</h3>
+    <ul>
+      <li>viewMode is optional <a
+              href='https://github.com/Eonasdan/tempus-dominus/issues/2550'>#2550</a></li>
+    </ul>
+    <h3>New</h3>
+    <ul>
+      <li>Introduced a simple overridable function parseInput  <a
+              href='https://github.com/Eonasdan/tempus-dominus/issues/2552'>#2552</a></li>
+    </ul>
     <h2>6-beta5</h2>
     <h3>Bug Fixes</h3>
     <ul>
