+++ src/docs/partials/examples/jquery.html
@@ -20,7 +20,7 @@
   </section>
 
   <section>
-    <h3 id='simple-setup'>Simple Setup <a class='anchorjs-link ' aria-label='Anchor' href='#simple-setup'><i
+    <h3 id='simple-setup'>Simple Setup <a class='anchor-link ' aria-label='Anchor' href='#simple-setup'><i
       class='fas fa-anchor' aria-hidden="true"></i></a> <small class='show-code' title='View Code'><i class='fas fa-code' aria-hidden="true"></i></small>
       <a aria-label='stackblitz' href='https://stackblitz.com/edit/tempus-dominus-v6-simple-jquery-setup?file=index.html'
          title='Stackblitz example'>
