+++ src/docs/partials/namespace/errors.html
@@ -11,21 +11,21 @@
   </p>
 
   <section>
-    <h2 id='unexpectedOption'>unexpectedOption <small>(code: 1)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='unexpectedOption'>unexpectedOption <small>(code: 1)</small><a class='anchor-link ' aria-label='Anchor'
                                                                           href='#unexpectedOption'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
       An error indicating that a key in the options object is invalid.
     </p>
 
-    <h2 id='unexpectedOptions'>unexpectedOptions <small>(code: 1)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='unexpectedOptions'>unexpectedOptions <small>(code: 1)</small><a class='anchor-link ' aria-label='Anchor'
                                                                             href='#unexpectedOptions'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
       An error indicating that one more keys in the options object is invalid.
     </p>
 
-    <h2 id='unexpectedOptionValue'>unexpectedOptionValue <small>(code: 2)</small><a class='anchorjs-link '
+    <h2 id='unexpectedOptionValue'>unexpectedOptionValue <small>(code: 2)</small><a class='anchor-link '
                                                                                     aria-label='Anchor'
                                                                                     href='#unexpectedOptionValue'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
@@ -34,7 +34,7 @@
       which only supports 'top', 'bottom', 'default'.
     </p>
 
-    <h2 id='typeMismatch'>typeMismatch <small>(code: 3)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='typeMismatch'>typeMismatch <small>(code: 3)</small><a class='anchor-link ' aria-label='Anchor'
                                                                   href='#typeMismatch'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
@@ -43,7 +43,7 @@
       supports true or false.
     </p>
 
-    <h2 id='numbersOutOfRage'>numbersOutOfRage <small>(code: 4)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='numbersOutOfRage'>numbersOutOfRage <small>(code: 4)</small><a class='anchor-link ' aria-label='Anchor'
                                                                           href='#numbersOutOfRage'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
@@ -51,7 +51,7 @@
       For example <code>restrictions.daysOfWeekDisabled</code> excepts a value between 0 and 6.
     </p>
 
-    <h2 id='failedToParseDate'>failedToParseDate <small>(code: 5)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='failedToParseDate'>failedToParseDate <small>(code: 5)</small><a class='anchor-link ' aria-label='Anchor'
                                                                             href='#failedToParseDate'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
@@ -59,14 +59,14 @@
       the option was an invalide string or an invalid Date object.
     </p>
 
-    <h2 id='mustProvideElement'>mustProvideElement <small>(code: 6)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='mustProvideElement'>mustProvideElement <small>(code: 6)</small><a class='anchor-link ' aria-label='Anchor'
                                                                               href='#mustProvideElement'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
       An error when an element to attach to was not provided in the constructor.
     </p>
 
-    <h2 id='subscribeMismatch'>subscribeMismatch <small>(code: 7)</small><a class='anchorjs-link ' aria-label='Anchor'
+    <h2 id='subscribeMismatch'>subscribeMismatch <small>(code: 7)</small><a class='anchor-link ' aria-label='Anchor'
                                                                             href='#subscribeMismatch'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
@@ -74,7 +74,7 @@
       the same number of callbacks. E.g., subscribe([1,2], [1])
     </p>
 
-    <h2 id='conflictingConfiguration'>conflictingConfiguration <small>(code: 8)</small><a class='anchorjs-link '
+    <h2 id='conflictingConfiguration'>conflictingConfiguration <small>(code: 8)</small><a class='anchor-link '
                                                                                           aria-label='Anchor'
                                                                                           href='#conflictingConfiguration'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
@@ -82,7 +82,7 @@
       The configuration has conflicting rules e.g. minDate is after maxDate
     </p>
 
-    <h2 id='dateString'>dateString<a class='anchorjs-link ' aria-label='Anchor' href='#dateString'><i
+    <h2 id='dateString'>dateString<a class='anchor-link ' aria-label='Anchor' href='#dateString'><i
       class="fa-solid fa-anchor" aria-hidden="true"></i></a></h2>
     <p>
       Logs a warning if a date option value is provided as a string, instead of
@@ -94,7 +94,7 @@
     <h2>Error Messages</h2>
     <div class='row'>
       <div class='col offset-1'>
-        <h3 id='failedToSetInvalidDate'>failedToSetInvalidDate<a class='anchorjs-link ' aria-label='Anchor'
+        <h3 id='failedToSetInvalidDate'>failedToSetInvalidDate<a class='anchor-link ' aria-label='Anchor'
                                                                  href='#failedToSetInvalidDate'><i
           class="fa-solid fa-anchor" aria-hidden="true"></i></a></h3>
         <p>
@@ -103,7 +103,7 @@
           validation.
         </p>
 
-        <h3 id='failedToParseInput'>failedToParseInput<a class='anchorjs-link ' aria-label='Anchor'
+        <h3 id='failedToParseInput'>failedToParseInput<a class='anchor-link ' aria-label='Anchor'
                                                          href='#failedToParseInput'><i
           class="fa-solid fa-anchor" aria-hidden="true"></i></a></h3>
         <p>
