+++ src/docs/partials/namespace/events.html
@@ -16,7 +16,7 @@ interface BaseEvent {
 }
     </code>
   </pre>
-    <h2 id="change">change.td <a class='anchorjs-link ' aria-label='Anchor' href='#change'><i
+    <h2 id="change">change.td <a class='anchor-link ' aria-label='Anchor' href='#change'><i
       class='fas fa-anchor' aria-hidden="true"></i></a></h2>
     <p>
       Emit when the date selection is changed.
@@ -31,7 +31,7 @@ interface ChangeEvent extends BaseEvent {
 }
     </code>
   </pre>
-    <h2 id="update">update.td <a class='anchorjs-link ' aria-label='Anchor' href='#update'><i
+    <h2 id="update">update.td <a class='anchor-link ' aria-label='Anchor' href='#update'><i
       class='fas fa-anchor' aria-hidden="true"></i></a></h2>
     <p>
       Emits when the view changes for example from month view to the year view.
@@ -43,7 +43,7 @@ interface ViewUpdateEvent extends BaseEvent {
 }
     </code>
   </pre>
-    <h2 id="error">error.td <a class='anchorjs-link ' aria-label='Anchor' href='#error'><i
+    <h2 id="error">error.td <a class='anchor-link ' aria-label='Anchor' href='#error'><i
       class='fas fa-anchor' aria-hidden="true"></i></a></h2>
     <p>
       Emits when a selected date or value from the input field fails to meet the provided validation rules.
@@ -57,12 +57,12 @@ interface FailEvent extends BaseEvent {
 }
     </code>
   </pre>
-    <h2 id="show">show.td <a class='anchorjs-link ' aria-label='Anchor' href='#show'><i
+    <h2 id="show">show.td <a class='anchor-link ' aria-label='Anchor' href='#show'><i
       class='fas fa-anchor' aria-hidden="true"></i></a></h2>
     <p>
       Emits when then picker widget is displayed.
     </p>
-    <h2 id="hide">hide.td <a class='anchorjs-link ' aria-label='Anchor' href='#hide'><i
+    <h2 id="hide">hide.td <a class='anchor-link ' aria-label='Anchor' href='#hide'><i
       class='fas fa-anchor' aria-hidden="true"></i></a></h2>
     <p>
       Emits when the picker widget is hidden.
