+++ src/docs/templates/index.html
@@ -16,7 +16,7 @@
                         <a href="/6/installing.html" class="btn btn-lg btn-outline-secondary mb-3" onclick="ga('send', 'event', 'Jumbotron actions', 'Download', 'Download 6');">Download</a>
                     </div>
                     <p class="text-muted mb-0">
-                        Currently <strong>v6.0.0-beta5.1</strong>
+                        Currently <strong>v6.0.0-beta7</strong>
                         <span class="px-1">&middot;</span>
                         <a href="https://getdatepicker.com/5-4/" class="link-secondary">v5 docs</a>
                     </p>
