+++ src/js/display/calendar/year-display.ts
@@ -1,11 +1,11 @@
 import { DateTime, Unit } from '../../datetime';
 import Namespace from '../../utilities/namespace';
 import Dates from '../../dates';
-import { OptionsStore } from '../../utilities/options';
 import Validation from '../../validation';
 import { Paint } from '../index';
 import { serviceLocator } from '../../utilities/service-locator';
 import ActionTypes from '../../utilities/action-types';
+import {OptionsStore} from "../../utilities/optionsStore";
 
 /**
  * Creates and updates the grid for `year`
