+++ src/js/display/time/minute-display.ts
@@ -1,10 +1,10 @@
 import { Unit } from '../../datetime';
 import Namespace from '../../utilities/namespace';
-import { OptionsStore } from '../../utilities/options';
 import Validation from '../../validation';
 import { serviceLocator } from '../../utilities/service-locator';
 import { Paint } from '../index';
 import ActionTypes from '../../utilities/action-types';
+import {OptionsStore} from "../../utilities/optionsStore";
 
 /**
  * Creates and updates the grid for `minutes`
