+++ src/js/jQuery-provider.js
@@ -2,7 +2,7 @@
 /*global $ */
 
 /*!
-  * Tempus Dominus v6.0.0-beta5.1 (https://getdatepicker.com/)
+  * Tempus Dominus v6.0.0-beta7 (https://getdatepicker.com/)
   * Copyright 2013-2021 [object Object]
   * Licensed under MIT (https://github.com/Eonasdan/tempus-dominus/blob/master/LICENSE)
   */
