+++ src/js/utilities/errors.ts
@@ -86,7 +86,7 @@ export class ErrorMessages {
 
   /**
    * Throws an error when a value for a date options couldn't be parsed. Either
-   * the option was an invalide string or an invalid Date object.
+   * the option was an invalid string or an invalid Date object.
    * @param optionName
    * @param date
    * @param soft If true, logs a warning instead of an error.
